The program now compiles and loads projects to and from disk. Components are now static variables. Added config file with macros. Scene window will now not move when it's not focused. Some refactoring to prevent unintended warnings

+++ CShard/CShard/src/device/graphics/GMesh.cpp
@@ -40,7 +40,7 @@ GMesh::GMesh(std::string filepath) : name(filepath)
 					attrib.normals[3 * index.normal_index + 2]
 				};
 	        vertices.push_back(vertex);
-	        indices.push_back(indices.size());
+	        indices.push_back((uint32_t)indices.size());
 	    }
 	}
 }
